% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{append_api_data_to_csv}
\alias{append_api_data_to_csv}
\title{Dump data to csv}
\usage{
append_api_data_to_csv(filename)
}
\arguments{
\item{filename}{\code{character} filename for resulting csv, should include file extension}
}
\description{
Dumps data from Warsaw queue api from all offices to csv file
}
