% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_available_offices}
\alias{get_available_offices}
\title{Get available offices}
\usage{
get_available_offices()
}
\value{
A \code{character} vector of names of offices in Warsaw
}
\description{
A function, that returns available office names to pass down to other methods.
}
