% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_available_queues}
\alias{get_available_queues}
\title{get_available_queues}
\usage{
get_available_queues(office_name)
}
\arguments{
\item{office_name}{\code{character} acronym of office in Warsaw.
You can get a list of possible values using \code{\link[kolejkeR]{get_available_offices}} function.}
}
\description{
get_available_queues
}
