% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_waiting_time_verbose}
\alias{get_waiting_time_verbose}
\alias{get_open_counters_verbose}
\alias{get_current_ticket_number_verbose}
\alias{get_number_of_people_verbose}
\title{Get specific data verbosely}
\usage{
get_waiting_time_verbose(office_name, queue_name, language = "en")

get_open_counters_verbose(office_name, queue_name, language = "en")

get_current_ticket_number_verbose(office_name, queue_name,
  language = "en")

get_number_of_people_verbose(office_name, queue_name, language = "en")
}
\arguments{
\item{office_name}{\code{character} acronym of office in Warsaw.

 You can get a list of possible values using \code{\link[kolejkeR]{get_available_offices}} function.}

\item{queue_name}{A \code{character} describing a queue we are interested in.
You can get a list of possible values using \code{\link[kolejkeR]{get_available_queues}} function.}

\item{language}{A \code{character}. Only two languages supported: English (\code{"en"}) and Polish (\code{"pl"}).}
}
\value{
A \code{character} vector of the same length as \code{queue_name} with each element in format depending on the called function and the variable \code{language}. Below we assume, that \code{language} variable is default.

If \code{get_waiting_time_verbose} is called: 

"Waiting time for <queue name> is x minutes".

If \code{get_open_encounters_verbose} is called: 

"There are x open encounters for <queue name>".

If \code{get_current_number_verbose} is called: 

"Current ticket number is x"

If \code{get_number_of_people_verbose} is called: 

"There are x people in <queue name>"
}
\description{
Several functions to get specific data, such as waiting time, open encounters, current ticket number and
amount of people in a set of specified queues in specified office.
}
\section{Functions}{
\itemize{
\item \code{get_waiting_time_verbose}: Returns expected time to be served.

\item \code{get_open_counters_verbose}: Returns amount of opened encounters.

\item \code{get_current_ticket_number_verbose}: Returns current ticket number.

\item \code{get_number_of_people_verbose}: Returns number of people waiting in specified queue.
}}

\examples{
office <- get_available_offices()[1]
queue <- get_available_queues(office)

get_waiting_time_verbose(office, queue)

get_open_counters_verbose(office, queue)

get_current_ticket_number_verbose(office, queue)

get_number_of_people_verbose(office, queue)
}
\seealso{
\code{\link[kolejkeR]{get_waiting_time}} and others to extract data directly.
}
