% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_waiting_time}
\alias{get_waiting_time}
\alias{get_open_counters}
\alias{get_current_ticket_number}
\alias{get_number_of_people}
\title{Get specific data directly}
\usage{
get_waiting_time(office_name, queue_name)

get_open_counters(office_name, queue_name)

get_current_ticket_number(office_name, queue_name)

get_number_of_people(office_name, queue_name)
}
\arguments{
\item{office_name}{\code{character} acronym of office in Warsaw.

 You can get a list of possible values using \code{\link[kolejkeR]{get_available_offices}} function.}

\item{queue_name}{A \code{character} vector describing the queues we are interested in.
You can get a list of possible values using \code{\link[kolejkeR]{get_available_queues}} function.}
}
\value{
A \code{character} vector (unless specified diffrently below) of the same length as \code{queue_name}, containing the information dependent on the called function.

If \code{get_waiting_time} is called: A \code{numeric} vector with estimated time of waiting in the queues, in minutes.

If \code{get_open_encounters} is called - A \code{numeric} vector with the amounts of opened encounters servicing the queues.

If \code{get_current_ticket_number} is called - the current ticket identifiers in the queues.

If \code{get_number_of_people} is called - A \code{numeric} vector with the amounts of people waiting in the queues.
}
\description{
Several functions to get specific data, such as waiting time, open encounters, current ticket number and
amount of people in a set of specific queues in specified office.
}
\section{Functions}{
\itemize{
\item \code{get_waiting_time}: Returns expected time to be served.

\item \code{get_open_counters}: Returns amount of opened encounters.

\item \code{get_current_ticket_number}: Returns the identifier of the current ticket.

\item \code{get_number_of_people}: Returns amount of people waiting in the queue.
}}

\examples{
office <- get_available_offices()[1]
queue <- get_available_queues(office)

get_waiting_time(office, queue)

get_open_counters(office, queue)

get_current_ticket_number(office, queue)

get_number_of_people(office, queue)
}
\seealso{
\code{\link[kolejkeR]{get_waiting_time_verbose}} and others for a more verbose output.
}
