% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetQueue.R
\name{GetQueue}
\alias{GetQueue}
\title{function GetQueue}
\usage{
GetQueue(office_acronym = "UD_Ochota")
}
\arguments{
\item{office_acronym}{\code{character} acronym of office in Warsaw}
}
\description{
function GetQueue
}
